requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.23.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pyyaml>=6.0
pillow>=9.0.0
